Refactor views directory names
