+++ src/index.tsx
@@ -5,11 +5,11 @@ import reportWebVitals from './reportWebVitals';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import { ChakraProvider } from '@chakra-ui/react';
 import { theme } from './theme/theme';
-import { LandingPage } from './views/LandingPage';
-import SignIn from './views/Authentication/SignIn';
-import SignUp from './views/Authentication/SignUp';
-import ForgotPassword from './views/Authentication/ForgotPassword';
-import ResetPassword from './views/Authentication/ResetPassword';
+import { LandingPage } from './views/landingpage';
+import SignIn from './views/authentication/SignIn';
+import SignUp from './views/authentication/SignUp';
+import ForgotPassword from './views/authentication/ForgotPassword';
+import ResetPassword from './views/authentication/ResetPassword';
 import NotFound from './views/NotFound';
 import Dashboard from './views/dashbaord/dashboard';
 import { Overview } from './views/dashbaord/overview';
